pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-FlyWise"
version = "0.1.0"
description = "Flywise environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.1",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
train = [
    "accelerate>=0.34.0",
    "datasets>=3.0.0",
    "openai>=1.40.0",
    "peft>=0.13.0",
    "torch>=2.1.0",
    "transformers>=4.46.0",
    "trl>=0.16.0",
]
inference = [
    "openai>=1.40.0",
    "peft>=0.13.0",
    "torch>=2.1.0",
    "transformers>=4.46.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m FlyWise.server.app
server = "FlyWise.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["FlyWise", "FlyWise.server"]
package-dir = { "FlyWise" = ".", "FlyWise.server" = "server" }